Clean up navigation states
Rename to use "go_to" as a self-explanatory prefix
Add teleportation navigation simulation fallback
