--- common_states/states/common_states/go_to_named_poi/statemachine.json
@@ -1,7 +1,0 @@
-{
-    "creation_time": "2020-02-15 02:35:41", 
-    "last_update": "2020-02-17 18:51:26", 
-    "root_state_storage_id": "go_to_named_poi_HGLHWO", 
-    "state_machine_version": null, 
-    "used_rafcon_version": "0.14.1"
-}
