+++ common_states/states/common_states/go_to_named_pose/statemachine.json
@@ -1,0 +1,7 @@
+{
+    "creation_time": "2020-02-15 02:35:41", 
+    "last_update": "2020-06-11 04:42:02", 
+    "root_state_storage_id": "go_to_named_pose_HGLHWO", 
+    "state_machine_version": null, 
+    "used_rafcon_version": "0.14.2"
+}
