+++ common_states/states/common_states/go_to_pose/go_to_pose_NXAMAO/core_data.json
@@ -15,7 +15,7 @@
             "name": "pose"
         }
     }, 
-    "name": "nav_goal", 
+    "name": "go_to_pose", 
     "outcomes": {
         "-2": {
             "__jsonqualname__": "rafcon.core.state_elements.logical_port.Outcome",
