+++ common_states/states/common_states/go_to_pose/go_to_pose_NXAMAO/meta_data.json
